# Optimization: HostWarmupMiddleware (JIT profile fix and async throughput)