--- a/src/WebJobs.Script.WebHost/Middleware/HostWarmupMiddleware.cs
+++ b/src/WebJobs.Script.WebHost/Middleware/HostWarmupMiddleware.cs
@@ -22,6 +22,10 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Middleware
         private readonly IScriptHostManager _hostManager;
         private readonly ILogger _logger;
         private string _assemblyLocalPath;
+        private volatile bool _jitTraceHasRun;
+
+        private static readonly PathString _warmupRoutePath = new PathString($"/api/{WarmUpConstants.FunctionName}");
+        private static readonly PathString _warmupRouteAlternatePath = new PathString($"/api/{WarmUpConstants.AlternateRoute}");
 
         public HostWarmupMiddleware(RequestDelegate next, IScriptWebHostEnvironment webHostEnvironment, IEnvironment environment, IScriptHostManager hostManager, ILogger<HostWarmupMiddleware> logger)
         {
@@ -33,20 +37,30 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Middleware
             _assemblyLocalPath = Path.GetDirectoryName(new Uri(typeof(HostWarmupMiddleware).Assembly.Location).LocalPath);
         }
 
-        public async Task Invoke(HttpContext httpContext)
+        public Task Invoke(HttpContext httpContext)
         {
             if (IsWarmUpRequest(httpContext.Request, _webHostEnvironment.InStandbyMode, _environment))
             {
-                PreJitPrepare(WarmUpConstants.JitTraceFileName);
-                if (_environment.IsLinuxConsumption())
-                {
-                    PreJitPrepare(WarmUpConstants.LinuxJitTraceFileName);
-                }
-
-                ReadRuntimeAssemblyFiles();
-
-                await WarmUp(httpContext.Request);
+                return WarmupInvoke(httpContext);
             }
+
+            return _next.Invoke(httpContext);
+        }
+
+        /// <summary>
+        /// This is so we only pay the async overhead while in the warmup path, but not for primary runtime.
+        /// </summary>
+        public async Task WarmupInvoke(HttpContext httpContext)
+        {
+            PreJitPrepare(WarmUpConstants.JitTraceFileName);
+            if (_environment.IsLinuxConsumption())
+            {
+                PreJitPrepare(WarmUpConstants.LinuxJitTraceFileName);
+            }
+
+            ReadRuntimeAssemblyFiles();
+
+            await WarmUp(httpContext.Request);
 
             await _next.Invoke(httpContext);
         }
@@ -96,6 +110,11 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Middleware
 
         private void PreJitPrepare(string jitTraceFileName)
         {
+            if (_jitTraceHasRun)
+            {
+                return;
+            }
+
             // This is to PreJIT all methods captured in coldstart.jittrace file to improve cold start time
             var path = Path.Combine(
                 _assemblyLocalPath,
@@ -112,6 +131,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Middleware
                 _logger.LogInformation(new EventId(100, "PreJit"),
                     $"PreJIT Successful prepares: {successfulPrepares}, Failed prepares: {failedPrepares} FileName = {jitTraceFileName}");
             }
+            _jitTraceHasRun = true;
         }
 
         public async Task WarmUp(HttpRequest request)
@@ -127,10 +147,9 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Middleware
 
         public static bool IsWarmUpRequest(HttpRequest request, bool inStandbyMode, IEnvironment environment)
         {
-            return inStandbyMode &&
-                ((environment.IsAppService() && request.IsAppServiceInternalRequest(environment)) || environment.IsLinuxConsumption()) &&
-                (request.Path.StartsWithSegments(new PathString($"/api/{WarmUpConstants.FunctionName}")) ||
-                request.Path.StartsWithSegments(new PathString($"/api/{WarmUpConstants.AlternateRoute}")));
+            return inStandbyMode
+                && ((environment.IsAppService() && request.IsAppServiceInternalRequest(environment)) || environment.IsLinuxConsumption())
+                && (request.Path.StartsWithSegments(_warmupRoutePath) || request.Path.StartsWithSegments(_warmupRouteAlternatePath));
         }
     }
 }